+++ src/main/java/comparator_comparable/Person.java
@@ -1,0 +1,18 @@
+package comparator_comparable;
+
+class Person implements Comparable<Person> {
+    String name;
+    int age;
+
+    public Person(String name, int age) {
+        super();
+        this.name = name;
+        this.age = age;
+    }
+
+    //사람의 정렬 기준은 나이
+    @Override
+    public int compareTo(Person o) {
+        return this.age-o.age;
+    }
+}
